+++ src/views/ErrorPage/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './404Page';
