+++ src/views/ErrorPage/styles.ts
@@ -1,0 +1,26 @@
+import styled from '@emotion/styled';
+
+export const Root = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 22px;
+
+  width: 100%;
+  height: 100vh;
+`;
+export const ErrorText = styled.p`
+  margin-bottom: 26px;
+
+  color: #FCFCFC;
+
+  font-size: 48px;
+  font-weight: 600;
+  line-height: 150%;
+  letter-spacing: -0.96px;
+`;
+export const CodeText = styled.div`
+  display: flex;
+  align-items: center;
+`;
